Move DatabaseDetails to pages directory
