+++ assets/js/pages/DatabaseDetails/DatabaseDetails.jsx
