+++ assets/js/pages/DatabaseDetails/DatabaseDetails.stories.jsx
