+++ assets/js/pages/DatabaseDetails/index.js
